+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/EngineSession.kt
@@ -494,6 +494,7 @@ abstract class EngineSession(
     /**
      * Represents settings options for cookie banner handling.
      */
+    @Suppress("MagicNumber")
     enum class CookieBannerHandlingMode(val mode: Int) {
         /**
          * The feature is turned off and cookie banners are not handled
@@ -509,6 +510,11 @@ abstract class EngineSession(
          * Reject cookies if possible. If rejecting is not possible, accept cookies
          */
         REJECT_OR_ACCEPT_ALL(2),
+
+        /**
+         * Detect cookie banners but do not handle them.
+         */
+        DETECT_ONLY(3),
     }
 
     /**